+++ src/components/Icons/iconVariables.ts
@@ -199,5 +199,11 @@ export const iconSizesArray = [
   "xlarge",
   "xxlarge",
   "xxxlarge",
+  "xxxxlarge",
+  "xxxxxlarge",
+  "2xlarge",
+  "3xlarge",
+  "4xlarge",
+  "5xlarge",
 ] as const;
 export const iconTypesArray = ["default", "breadcrumbs"] as const;